move scripts to src
