--- src/hw-2/index3-1.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-  <head>
-    <meta charset="utf-8">
-    <title>HW-3</title>
-  </head>
-  <body>
-      <script type="text/javascript" src="script3-1.js">
-      </script>
-  </body>
-</html>
