+++ src/index2.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+  <head>
+    <meta charset="utf-8" />
+    <title>HW-3</title>
+  </head>
+  <body>
+    <script type="text/javascript" src="script2.js"></script>
+  </body>
+</html>
