+++ src/script2.js
